requirements.txt
certifi==2024.2.2
chardet==5.2.0
click==8.1.7
dominate==2.9.1
Flask==3.0.2
Bootstrap-Flask==2.3.3
Flask-WTF==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
requests==2.31.0
urllib3==2.2.1
visitor==0.1.3
Werkzeug==3.0.1
WTForms==3.1.2
nh3==0.2.15
pandas==2.2.1
numpy==1.26.4
pyarrow==15.0.0
pillow==10.2.0